# backend/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"PromptBox API","version":"0.1.0"},"paths":{"/health/":{"get":{"summary":"Health","operationId":"health_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/validate-key":{"post":{"summary":"Validate Key","operationId":"validate_key_api_validate_key_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateKeyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/execute":{"post":{"summary":"Execute","operationId":"execute_api_execute_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/prompt":{"post":{"summary":"Prompt Batch","operationId":"prompt_batch_api_prompt_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptBatchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromptBatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/versions/":{"post":{"tags":["versions"],"summary":"Save Version","operationId":"save_version_api_versions__post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionCreateRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["versions"],"summary":"Get Versions","operationId":"get_versions_api_versions__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VersionResponse"},"title":"Response Get Versions Api Versions  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/versions/{version_id}":{"get":{"tags":["versions"],"summary":"Get Version","operationId":"get_version_api_versions__version_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["versions"],"summary":"Update Version","operationId":"update_version_api_versions__version_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionUpdateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["versions"],"summary":"Delete Version","operationId":"delete_version_api_versions__version_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/export/{version_id}":{"get":{"tags":["export"],"summary":"Export Version","operationId":"export_version_api_export__version_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}},{"name":"format","in":"query","required":true,"schema":{"type":"string","description":"Format to export (txt, md, json)","title":"Format"},"description":"Format to export (txt, md, json)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/templates/":{"get":{"tags":["templates"],"summary":"Get Templates","operationId":"get_templates_api_templates__get","parameters":[{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TemplateResponse"},"title":"Response Get Templates Api Templates  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/templates/{template_id}":{"get":{"tags":["templates"],"summary":"Get Template","operationId":"get_template_api_templates__template_id__get","parameters":[{"name":"template_id","in":"path","required":true,"schema":{"type":"integer","title":"Template Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemplateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Credentials":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"Credentials"},"ExecuteRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"model":{"type":"string","title":"Model"},"api_key":{"type":"string","title":"Api Key"}},"type":"object","required":["prompt","model","api_key"],"title":"ExecuteRequest"},"ExecuteResponse":{"properties":{"model":{"type":"string","title":"Model"},"response_text":{"type":"string","title":"Response Text"},"latency":{"type":"number","title":"Latency"}},"type":"object","required":["model","response_text","latency"],"title":"ExecuteResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PromptBatchRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"models":{"items":{"type":"string"},"type":"array","title":"Models"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"api_keys":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Api Keys"}},"type":"object","required":["prompt","models"],"title":"PromptBatchRequest"},"PromptBatchResponse":{"properties":{"responses":{"items":{"$ref":"#/components/schemas/ExecuteResponse"},"type":"array","title":"Responses"}},"type":"object","required":["responses"],"title":"PromptBatchResponse"},"TemplateResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"category":{"type":"string","title":"Category"},"content":{"type":"string","title":"Content"}},"type":"object","required":["id","name","category","content"],"title":"TemplateResponse"},"ValidateKeyRequest":{"properties":{"provider":{"type":"string","title":"Provider"},"api_key":{"type":"string","title":"Api Key"}},"type":"object","required":["provider","api_key"],"title":"ValidateKeyRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VersionCreateRequest":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"tag":{"anyOf":[{"type":"string","maxLength":32},{"type":"null"}],"title":"Tag"},"prompt_text":{"type":"string","minLength":1,"title":"Prompt Text"},"response_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Response Text"},"response_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Response Model"},"response_latency":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Response Latency"}},"type":"object","required":["name","prompt_text"],"title":"VersionCreateRequest"},"VersionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"tag":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tag"},"prompt_text":{"type":"string","title":"Prompt Text"},"response_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Response Text"},"response_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Response Model"},"response_latency":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Response Latency"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","prompt_text","created_at"],"title":"VersionResponse"},"VersionUpdateRequest":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"tag":{"anyOf":[{"type":"string","maxLength":32},{"type":"null"}],"title":"Tag"}},"type":"object","required":["name"],"title":"VersionUpdateRequest"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}